Rholang: Move substituteAndCharge methods to Substitute object.
