+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/CostAccountingReducerTest.scala
@@ -8,7 +8,13 @@ import coop.rchain.models.Var.VarInstance.{BoundVar, FreeVar}
 import coop.rchain.models._
 import coop.rchain.models.rholang.implicits._
 import coop.rchain.rholang.interpreter.Reduce.DebruijnInterpreter
-import coop.rchain.rholang.interpreter.accounting.{Chargeable, Cost, CostAccount, CostAccountingAlg, _}
+import coop.rchain.rholang.interpreter.accounting.{
+  Chargeable,
+  Cost,
+  CostAccount,
+  CostAccountingAlg,
+  _
+}
 import coop.rchain.rholang.interpreter.errors.OutOfPhlogistonsError
 import coop.rchain.rholang.interpreter.storage.implicits._
 import coop.rchain.rholang.interpreter.storage.{ChargingRSpace, ChargingRSpaceTest, TuplespaceAlg}
@@ -32,7 +38,7 @@ class CostAccountingReducerTest extends FlatSpec with Matchers with TripleEquals
     val termCost          = Chargeable[Par].cost(substTerm)
     val initCost          = CostAccount(1000)
     val costAlg           = CostAccountingAlg.unsafe[Coeval](initCost)
-    val res               = Reduce.substituteAndCharge[Par, Coeval](varTerm, 0, env, costAlg).attempt.value
+    val res               = Substitute.substituteAndCharge[Par, Coeval](varTerm, 0, env, costAlg).attempt.value
     assert(res === Right(substTerm))
     assert(costAlg.get().value.cost === (initCost.cost - Cost(termCost)))
   }
@@ -44,7 +50,7 @@ class CostAccountingReducerTest extends FlatSpec with Matchers with TripleEquals
     val originalTermCost  = Chargeable[Par].cost(varTerm)
     val initCost          = CostAccount(1000)
     val costAlg           = CostAccountingAlg.unsafe[Coeval](initCost)
-    val res               = Reduce.substituteAndCharge[Par, Coeval](varTerm, 0, env, costAlg).attempt.value
+    val res               = Substitute.substituteAndCharge[Par, Coeval](varTerm, 0, env, costAlg).attempt.value
     assert(res.isLeft)
     assert(costAlg.get().value.cost === (initCost.cost - Cost(originalTermCost)))
   }
